// react-app/src/misc/openapi.d.ts
export interface components {
    schemas: {
// ...
        CreateSectionRequest: {
            name: string;
            displayName: string;
            color: string;
        };
        DeleteNoteRequest: {
            /** Format: int64 */
            id: number;
        };
        DeleteSectionRequest: {
            /** Format: int64 */
            id: number;
        };
// ...
        ReorderSectionsRequest: {
            ids: number[];
        };
// ...
        SectionsResponse: {
            message?: string | null;
            sections: components["schemas"]["SectionItemResponse"][] | null;
        };
// ...
        UpdateSectionRequest: {
            /** Format: int64 */
            id: number;
            displayName: string;
            color: string;
        };
        UserResponse: {
            id?: string | null;
            email?: string | null;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}